+++ plugin_ide.ui/src/melnorme/lang/ide/ui/LangUIPlugin_Actual.java
@@ -33,6 +33,6 @@ public final class LangUIPlugin_Actual {
 		return new GoAutoEditStrategy(contentType, sourceViewer);
 	}
 	
-	public static final String DAEMON_TOOL_ConsoleName = "gocode log";
+	public static final String DAEMON_TOOL_ConsoleName = "Oracle/gocode log";
 	
 }