+++ documentation/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ### Goclipse 0.9.x
  * Fixed: Builder does not ignore folders with names stating with "_" or "."
+   * Now using "./..." pattern to list files to compile.
  * Fixed: Project wizard: only create "bin,pkg,src" folders if project is not in the GOPATH.
  
 ▶ Recommended/tested CDT version: 8.4.0